Rename jog_arm->moveit_jog_arm (#1663)
* Rename jog_arm->moveit_jog_arm

* Address @rhaschke's feedback

* apply clang-format
