--- moveit_experimental/jog_arm/config/spacenav_via_teleop_tools.yaml
@@ -1,36 +1,0 @@
-teleop:
-  cartesian_twist_command:
-    type: topic
-    message_type: geometry_msgs/TwistStamped
-    topic_name: jog_arm_server/delta_jog_cmds
-    axis_mappings:
-      -
-        axis: 0
-        target: twist.linear.x
-        scale: 1
-        offset: 0
-      -
-        axis: 1
-        target: twist.linear.y
-        scale: 1
-        offset: 0
-      -
-        axis: 2
-        target: twist.linear.z
-        scale: 1
-        offset: 0
-      -
-        axis: 3
-        target: twist.angular.x
-        scale: 1
-        offset: 0
-      -
-        axis: 4
-        target: twist.angular.y
-        scale: 1
-        offset: 0
-      -
-        axis: 5
-        target: twist.angular.z
-        scale: 1
-        offset: 0
