+++ moveit_experimental/moveit_jog_arm/launch/spacenav_teleop_tools.launch
@@ -1,0 +1,27 @@
+<launch>
+  <!-- The package has 2 examples of converting controller commands to jog commands.
+       This is the higher-latency, Python-based teleop_tools example.
+       You can modify it in your own package for your own type of gamepad or joystick.
+       We do plan to accept pull requests of config files for other controller types using this method.
+  -->
+
+  <!-- This node publishes commands from the controller -->
+  <node name="spacenav_node" pkg="spacenav_node" type="spacenav_node" />
+
+  <!-- This node does the jogging calculations -->
+  <node name="jog_server" pkg="moveit_jog_arm" type="jog_server" output="screen" >
+    <param name="parameter_ns" type="string" value="jog_server" />
+    <rosparam command="load" file="$(find moveit_jog_arm)/config/ur_simulated_config.yaml" />
+  </node>
+
+  <!-- teleop_tools translates a joystick command into a twist message -->
+  <arg name="cmd_vel" default="spacenav/joy" />
+  <arg name="teleop_config" default="$(find moveit_jog_arm)/config/spacenav_via_teleop_tools.yaml" />
+  <rosparam file="$(arg teleop_config)" command="load" />
+
+  <node pkg="joy_teleop" type="joy_teleop.py" name="joy_teleop">
+    <remap from="cmd_vel" to="$(arg cmd_vel)" />
+    <remap from="joy" to="spacenav/joy" />
+  </node>
+
+</launch>
