+++ moveit_experimental/moveit_jog_arm/package.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>moveit_jog_arm</name>
+  <version>0.0.3</version>
+
+  <description>Provides real-time manipulator Cartesian jogging.</description>
+
+  <maintainer email="blakeanderson@utexas.edu">Blake Anderson</maintainer>
+  <maintainer email="andyz@utexas.edu">Andy Zelenak</maintainer>
+
+  <license>BSD 3-Clause</license>
+
+  <url type="website">https://ros-planning.github.io/moveit_tutorials</url>
+
+  <author>Brian O'Neil</author>
+  <author email="andyz@utexas.edu">Andy Zelenak</author>
+  <author>Blake Anderson</author>
+  <author email="alex@machinekoder.com">Alexander Rössler</author>
+
+  <buildtool_depend>catkin</buildtool_depend>
+
+  <depend>control_msgs</depend>
+  <depend>geometry_msgs</depend>
+  <depend>moveit_ros_planning_interface</depend>
+  <depend>rosparam_shortcuts</depend>
+  <depend>tf2_ros</depend>
+
+  <exec_depend>joy_teleop</exec_depend>
+
+  <test_depend>rostest</test_depend>
+  <test_depend>ros_pytest</test_depend>
+  <test_depend>moveit_resources</test_depend>
+
+  <export>
+  </export>
+
+</package>
